Add round log, line and ray drawing, projectiles

+++ Components/Item.cs
@@ -16,7 +16,7 @@ namespace TFT_Engine.Components
         public virtual void OnTick() { }
         public virtual void OnEnd() { }
         public virtual void OnSpecialMove() { }
-        public virtual void OnHit(Character attacker, DamageType damageType, bool isSpecial, bool isCrit, int amount) { }
+        public virtual void OnHit(Character attacker, DamageType damageType, bool isSpecial, bool isCrit, float amount) { }
         public virtual void OnAttack(Character target) { }
         public virtual void OnManaChange(int ManaChangeAmount) { }
         public virtual int OnDamageCalculation(int damage) { return damage; }
